Refatorando para considerar regras de autorização

+++ backend/FinanceGO.Application/Commands/ReceitaCommands/DeleteReceita/DeleteReceitaCommandHandler.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using FinanceGO.Core.Authorization;
 using FinanceGO.Core.Repositories.ReceitaRepositories;
 using FinanceGO.Core.Results;
 using MediatR;
@@ -13,16 +14,21 @@ namespace FinanceGO.Application.Commands.ReceitaCommands.DeleteReceita
     {
         private readonly IReceitaQueryRepository _queryRepository;
         private readonly IReceitaCommandRepository _commandRepository;
+        private readonly IMesmoUsuarioAuthorizationRequirement _requirement;
 
-        public DeleteReceitaCommandHandler(IReceitaQueryRepository queryRepository, IReceitaCommandRepository commandRepository)
+        public DeleteReceitaCommandHandler(IReceitaQueryRepository queryRepository, IReceitaCommandRepository commandRepository, IMesmoUsuarioAuthorizationRequirement requirement)
         {
             _queryRepository = queryRepository;
             _commandRepository = commandRepository;
+            _requirement = requirement;
         }
         public async Task<Result> Handle(DeleteReceitaCommand request, CancellationToken cancellationToken)
         {
             var receitaASerDeletada = await _queryRepository.GetReceitaByIdAsync(request.Id);
             if(receitaASerDeletada == null) return new RegistroNaoEncontradoResult();
+
+            var usuarioAutorizado = _requirement.VerificarReceitaMesmoUsuario(receitaASerDeletada);
+            if(!usuarioAutorizado) return new UsuarioNaoAutorizadoResult();
 
             await _commandRepository.DeleteReceitaAsync(receitaASerDeletada);
             return new DeletadoComSucessoResult();
